ai-service/requirements.txt
ultralytics
opencv-contrib-python
python-dotenv
PyYAML
numpy
torch
torchvision
torchaudio
fastapi
uvicorn[standard]